requirements.txt
requests
pandas
websockets
loguru
python-dotenv